BUG:222354 I have a nasty feeling I remove this feature on purpose, but I can't find the details, so lets just get this back in.
Thanks to Andreas Neustifter for the patch.

svn path=/trunk/extragear/graphics/kphotoalbum/; revision=1137915

+++ Browser/BrowserWidget.cpp
@@ -167,7 +167,7 @@ void Browser::BrowserWidget::load( const QString& category, const QString& value
     info.addAnd( category, value );
 
     DB::MediaCount counts = DB::ImageDB::instance()->count( info );
-    bool loadImages = (counts.total() < static_cast<uint>(Settings::SettingsData::instance()->autoShowThumbnailView()));
+    bool loadImages = (counts.total() < Settings::SettingsData::instance()->autoShowThumbnailView());
     if ( Utilities::ctrlKeyDown() ) loadImages = !loadImages;
 
     if ( loadImages )
